// eventos/apps/frontend/src/core/evento/model/Evento.ts
import Convidado from "./Convidado";

export default interface Evento{
    id: string;
    alias: string; // apelido
    senha: string;
    nome: string;
    data: Date;
    local: string;
    descricao: string;
    imagem: string;
    imagemBackground: string;
    publicoEsperado: number;
    convidados: Convidado[];   
}